Refactor ID types from size_t to long in various modules
- Updated function signatures in armes.h, bibelots.h, objets.h to use long instead of size_t for ID parameters.
- Modified implementations in armes.c, bibelots.c, objets.c to accommodate the new ID type.
- Adjusted print statements in display.c to reflect the change in ID type.
- Refactored bestiary initialization in creatures.c to handle groups and creatures with long IDs.
- Updated configuration loading functions to handle long IDs for creatures and groups.
- Ensured proper memory management for new structures in bestiary.
- Adjusted save/load functionality in sauvegarde.c to handle long IDs for equipped weapons.

+++ code/src/armes.c
@@ -117,7 +117,7 @@ void afficherArmes(Arsenal *arsenal) {
     printf("\n=== Arsenal disponible ===\n");
     for (size_t i = 0; i < arsenal->longueur; i++) {
         Arme *a = arsenal->armes[i];
-        printf("[%zu] %s (id=%zu) (ATK %d-%d | Coût O2: %d | DEF+%d) (Rareté: %d) / desc: '%s'\n",
+        printf("[%zu] %s (id=%ld) (ATK %d-%d | Coût O2: %d | DEF+%d) (Rareté: %d) / desc: '%s'\n",
                i, a->nom, a->id, a->attaque_min, a->attaque_max, a->cout_oxygene,
                a->bonus_defense, a->rarete, a->description);
         printListeAction(a->listeAction, "\t");
@@ -171,14 +171,14 @@ Arme *duplicateArme(Arme *a) {
 }
 
 
-int ajouterArme(Arsenal *modal, Arsenal *arsenal, size_t id_arme) {
-    if (!modal || !arsenal || id_arme >= modal->longueur) {
+int ajouterArme(Arsenal *modal, Arsenal *arsenal, long id_arme) {
+    if (!modal || !arsenal || (size_t) id_arme >= modal->longueur) {
         fprintf(stderr, "Erreur: ajouterArme(): paramètres invalides\n");
         return EXIT_FAILURE;
     }
 
-    if (id_arme > modal->longueur) {
-        fprintf(stderr, "Erreur: ajouterArme(): id_arme (%zu) hors limites (%zu)\n", id_arme, modal->longueur);
+    if ((size_t) id_arme > modal->longueur) {
+        fprintf(stderr, "Erreur: ajouterArme(): id_arme (%ld) hors limites (%zu)\n", id_arme, modal->longueur);
         return EXIT_FAILURE;
     }
 
@@ -206,8 +206,8 @@ int ajouterArme(Arsenal *modal, Arsenal *arsenal, size_t id_arme) {
     return EXIT_SUCCESS;
 }
 
-int equiperArme(Plongeur *joueur, size_t id_arme) {
-    if (!joueur || joueur->arsenal->longueur == 0 || id_arme >= joueur->arsenal->longueur) {
+int equiperArme(Plongeur *joueur, long id_arme) {
+    if (!joueur || joueur->arsenal->longueur == 0 || (size_t) id_arme >= joueur->arsenal->longueur) {
         fprintf(stderr, "Erreur: equiperArme(): paramètres invalides\n");
         return EXIT_FAILURE;
     }
